+++ domain/src/main/java/com/pickcar/auth/application/AuthService.java
@@ -28,7 +28,7 @@ public class AuthService {
     private final UserRepository userRepository;
     private final JwtProvider jwtProvider;
     private final PasswordEncoder passwordEncoder;
-    private final RefreshTokenRepository refreshTokenRepository;
+    private final TokenService tokenService;
 
     //TODO: 에러처리 하기
     public User getById(Long id) {
@@ -87,51 +87,8 @@ public class AuthService {
         );
 
         String refreshToken = jwtProvider.createRefreshToken(user.getId());
-        saveOrUpdateRefreshToken(user.getId(),refreshToken);
+        tokenService.saveOrUpdateRefreshToken(user.getId(),refreshToken);
         return new AuthResponse(accessToken,refreshToken);
-    }
-
-    @Transactional
-    public AuthResponse reissueTokens(String refreshToken){
-        //TODO : 변수명, 예외처리 수정하기
-        RefreshToken token = refreshTokenRepository.findByToken(refreshToken)
-                .orElseThrow(() -> new IllegalArgumentException("token이 없음"));
-
-        if(token.isExpired()){ //TODO: 없는 경우 예외처리 수정하기
-            throw new IllegalStateException("Refresh token has expired.");
-        }
-
-        User user = userRepository.findById(token.getUserId())
-                .orElseThrow(() -> new IllegalArgumentException("User not found."));
-
-        String newAccessToken = jwtProvider.createAccessToken(
-                user.getId(),
-                user.getInfo().getName(),
-                user.getRole().name()
-        );
-
-        String newRefreshToken = jwtProvider.createRefreshToken(user.getId());
-        saveOrUpdateRefreshToken(user.getId(), newRefreshToken);
-        return new AuthResponse(newAccessToken, newRefreshToken);
-    }
-
-    @Transactional
-    public void deleteByToken(String token){
-        refreshTokenRepository.deleteByToken(token);
-    }
-
-    private void saveOrUpdateRefreshToken(Long userId,String refreshToken){
-        //TODO: 해시화 해서 저장하기
-        LocalDateTime expiryDate = JwtUtils.calculateExpiryDate(JwtConstants.REFRESH_TOKEN_VALIDITY);
-        refreshTokenRepository.findByUserId(userId)
-                .ifPresentOrElse(
-                        token -> token.update(refreshToken,expiryDate),
-                        () -> refreshTokenRepository.save(RefreshToken.create(
-                                userId,
-                                refreshToken,
-                                JwtUtils.calculateExpiryDate(JwtConstants.REFRESH_TOKEN_VALIDITY)
-                        ))//TODO: 토큰 저장 실패 예외처리 하기
-                );
     }
 
     private UserInfo toUserInfoWithEncodedPassword(UserInfoRequest request) {
