+++ domain/src/main/java/com/pickcar/auth/presentation/AuthController.java
@@ -1,6 +1,7 @@
 package com.pickcar.auth.presentation;
 
 import com.pickcar.auth.application.AuthService;
+import com.pickcar.auth.application.TokenService;
 import com.pickcar.auth.presentation.dto.request.AuthRequest;
 import com.pickcar.auth.presentation.dto.request.UserInfoRequest;
 import com.pickcar.auth.presentation.dto.response.AccessTokenResponse;
@@ -24,6 +25,7 @@ import java.util.List;
 public class AuthController {
 
     private final AuthService authService;
+    private final TokenService tokenService;
 
     @PostMapping("/sign-up")
     @ResponseStatus(HttpStatus.CREATED)
@@ -56,7 +58,7 @@ public class AuthController {
 
         if (refreshToken != null && !refreshToken.isBlank()) {
             try {
-                authService.deleteByToken(refreshToken);
+                tokenService.deleteByToken(refreshToken);
             } catch (Exception e) {
                 log.warn("로그아웃 중 RT 삭제 실패: {}", e.getMessage());
             }
